+++ src/game/nlcore/common/game.ts
@@ -1,7 +1,7 @@
 import {Game} from "@core/game";
 import {GameState} from "@player/gameState";
 import {Storable, Namespace} from "../elements/persistent/storable";
-import {LiveGame} from "@core/liveGame";
+import {LiveGame} from "@core/game/liveGame";
 
 export {
     LiveGame,